convert class component to functoon
